Image-Text-Retrieval/requirements.txt
absl-py==1.4.0
cachetools==5.3.0
charset-normalizer==3.1.0
cycler==0.11.0
fonttools==4.38.0
ftfy==6.1.1
google-auth==2.17.3
google-auth-oauthlib==0.4.6
grpcio==1.54.0
idna==3.4
importlib-metadata==6.6.0
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
